Remove the BootstrapSingleton
